+++ climpred/tests/test_hindcast_prediction.py
@@ -16,15 +16,23 @@ from climpred.tutorial import load_dataset
 @pytest.fixture
 def initialized_ds():
     da = load_dataset('CESM-DP-SST')
-    da = da.sel(init=slice(1955, 2015))
     da = da - da.mean('init')
+    return da
+
+
+@pytest.fixture
+def initialized_ds_lead0():
+    da = load_dataset('CESM-DP-SST')
+    da = da - da.mean('init')
+    # Change to a lead-0 framework
+    da['init'] += 1
+    da['lead'] -= 1
     return da
 
 
 @pytest.fixture
 def initialized_da():
     da = load_dataset('CESM-DP-SST')['SST']
-    da = da.sel(init=slice(1955, 2015))
     da = da - da.mean('init')
     return da
 
@@ -46,8 +54,6 @@ def observations_da():
 @pytest.fixture
 def reconstruction_ds():
     da = load_dataset('FOSI-SST')
-    # same timeframe as DPLE
-    da = da.sel(time=slice(1955, 2015))
     da = da - da.mean('time')
     return da
 
@@ -55,8 +61,6 @@ def reconstruction_ds():
 @pytest.fixture
 def reconstruction_da():
     da = load_dataset('FOSI-SST')['SST']
-    # same timeframe as DPLE
-    da = da.sel(time=slice(1955, 2015))
     da = da - da.mean('time')
     return da
 
@@ -104,7 +108,7 @@ def test_compute_hindcast_less_m2r(initialized_da, reconstruction_da):
 @pytest.mark.parametrize('comparison', HINDCAST_COMPARISONS)
 def test_compute_hindcast(initialized_ds, reconstruction_ds, metric, comparison):
     """
-    Checks that compute reference works without breaking.
+    Checks that compute hindcast works without breaking.
     """
     res = (
         compute_hindcast(
@@ -115,6 +119,24 @@ def test_compute_hindcast(initialized_ds, reconstruction_ds, metric, comparison)
     )
     for var in res.data_vars:
         assert not res[var]
+
+
+@pytest.mark.parametrize('metric', PM_METRICS)
+@pytest.mark.parametrize('comparison', HINDCAST_COMPARISONS)
+def test_compute_hindcast_lead0_lead1(
+    initialized_ds, initialized_ds_lead0, reconstruction_ds, metric, comparison
+):
+    """
+    Checks that compute hindcast returns the same results with a lead-0 and lead-1
+    framework.
+    """
+    res1 = compute_hindcast(
+        initialized_ds, reconstruction_ds, metric=metric, comparison=comparison
+    )
+    res2 = compute_hindcast(
+        initialized_ds_lead0, reconstruction_ds, metric=metric, comparison=comparison
+    )
+    assert (res1.SST.values == res2.SST.values).all()
 
 
 @pytest.mark.parametrize('metric', PM_METRICS)
@@ -131,6 +153,19 @@ def test_persistence(initialized_da, reconstruction_da, metric):
 
 
 @pytest.mark.parametrize('metric', PM_METRICS)
+def test_persistence_lead0_lead1(
+    initialized_ds, initialized_ds_lead0, reconstruction_ds, metric
+):
+    """
+    Checks that compute persistence returns the same results with a lead-0 and lead-1
+    framework.
+    """
+    res1 = compute_persistence(initialized_ds, reconstruction_ds, metric=metric)
+    res2 = compute_persistence(initialized_ds_lead0, reconstruction_ds, metric=metric)
+    assert (res1.SST.values == res2.SST.values).all()
+
+
+@pytest.mark.parametrize('metric', PM_METRICS)
 @pytest.mark.parametrize('comparison', HINDCAST_COMPARISONS)
 def test_uninitialized(uninitialized_da, reconstruction_da, metric, comparison):
     """
@@ -138,11 +173,7 @@ def test_uninitialized(uninitialized_da, reconstruction_da, metric, comparison):
     """
     res = (
         compute_uninitialized(
-            uninitialized_da,
-            reconstruction_da,
-            metric=metric,
-            comparison=comparison,
-            nlags=5,
+            uninitialized_da, reconstruction_da, metric=metric, comparison=comparison
         )
         .isnull()
         .any()