+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 Changelog History
 =================
 
+climpred v1.0.2 (2019-07-##)
+============================
+
+Internals/Minor Fixes
+---------------------
+- Add `matplotlib` as a main dependency so that a direct pip installation works (:pr:`211`) `Riley X. Brady`_.
+
 climpred v1.0.1 (2019-07-04)
 ============================